OnlineJumpPointLocator/JumpPointLocator: added ::fringe, a collection to store forced neighbours or jump point successors of nodes returned by ::findJumpNode. Needed for multi-step jumping.
driver: ignoring cleanup routine for now (segfault under cygwin).

other changes: merged changes from trunk (mostly Makefile changes to compile
under cygwin and fixes for gcc344 warnings).

+++ branches/multijump/abstraction/MapLineAbstraction.cpp
@@ -95,9 +95,9 @@ void MapLineAbstraction::addNodes(graph *g)
 {
 	int count = abstractions.back()->getNumNodes();
 	//printf("Initial count: %d\n", count);
-	int xstep = pow(lineDistance,((abstractions.size()+1)/2));
-	int ystep = pow(lineDistance,((abstractions.size())/2));
-	int zstep = pow(lineDistance,((abstractions.size()-1)/2));
+	int xstep = (int)pow(lineDistance,((abstractions.size()+1)/2));
+	int ystep = (int)pow(lineDistance,((abstractions.size())/2));
+	int zstep = (int)pow(lineDistance,((abstractions.size()-1)/2));
 	graph *toAbstract = abstractions.back();
 	//printf("Size: %d.[%d] XStep: %d, YStep: %d, ZStep %d\n", abstractions.size(), lineDistance, xstep, ystep, zstep);
